+++ ariles/cmake/options_ros.cmake
@@ -3,5 +3,6 @@
 set(ARILES_VISITOR_msgpack      "OFF"   CACHE STRING "")
 set(ARILES_VISITOR_pugixml      "OFF"   CACHE STRING "")
 set(ARILES_VISITOR_rapidjson    "OFF"   CACHE STRING "")
+set(ARILES_VISITOR_nlohmann_json "OFF"  CACHE STRING "")
 set(ARILES_VISITOR_jsonnet      "OFF"   CACHE STRING "")
 set(ARILES_VISITOR_ros2param    "OFF"   CACHE STRING "")